fix(core): bound the numbers a team file may hold
Three of the landing review's findings, all about a person getting a
refusal that names the wrong thing or none at all.

The four session limits and an exit code had no ceiling, so a number
past what serde could hold reached the typed build and came back at the
root: "the schema passed but the typed team could not be built: invalid
type: floating point 1e+19", or, worse, "data did not match any variant
of untagged enum CriterionTemplateVerification" -- a generated Rust type
named at someone editing YAML. The session limits stop at the largest
integer JSON holds exactly, an exit code at what a process can return,
and both are now refused at their own pointer. The project plan already
recorded exit_code as an i32; the domain type agreed with the wire and
not with the plan, and now agrees with both.

A criterion library could hold two hundred criteria while a contract
takes a hundred exit criteria, so a library the validator accepted could
expand into a contract that was refused, a long way from the mistake.
A hundred, for the reason spec 3 gives for every other list: the refusal
prints what it refused.

And nothing held the integer normalisation that both new validators do:
deleting it from either left every test green, while a team file saying
wip_limit_per_agent: 2.0 -- which JSON Schema calls an integer -- would
have been refused at the root. Two tests now, one per validator.

Every ceiling in both schemas that a test can reach has one now: the
team's name, an agent's id and display name, a model id, an integration
branch, the rules lists, and the library's own length.

+++ crates/core/src/contract.rs
@@ -25,8 +25,9 @@ pub enum Verification {
     Command {
         /// Run inside the sandbox from the project root.
         command: String,
-        /// The exit code that counts as a pass; the schema's default is 0.
-        exit_code: i64,
+        /// The exit code that counts as a pass; the schema's default is 0. What a process can
+        /// return, which is what the schema now says too.
+        exit_code: i32,
         /// Text the standard output must contain.
         stdout_contains: Option<String>,
         /// Text the standard output must not contain.
